refactor: Move PieChart comp
